backend/file_rag/requirements.txt
# File RAG 系统依赖包

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# LangChain 核心
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# LLM 提供商
langchain-openai>=0.0.5
langchain-ollama>=0.1.0

# 向量数据库
langchain-milvus>=0.1.0
pymilvus>=2.3.0

# 文档处理
langchain-text-splitters>=0.0.1
pypdf>=3.17.0
pymupdf>=1.23.0
langchain-pymupdf4llm>=0.1.0

# Word 文档处理
docx2txt>=0.8
python-docx>=0.8.11

# 其他文档格式
pypandoc>=1.11
python-pptx>=0.6.21
openpyxl>=3.1.0
pandas>=2.0.0

# LangGraph
langgraph>=0.0.20

# 工具和实用程序
typing-extensions>=4.8.0

